Separate presentational component out of the Gauge component
