+++ frontend/src/metabase/static-viz/components/Gauge/Gauge.tsx
@@ -1,34 +1,25 @@
-import React, { Fragment } from "react";
-
+import React from "react";
 import { Pie } from "@visx/shape";
 import { Group } from "@visx/group";
-import type { PieArcDatum } from "@visx/shape/lib/shapes/Pie";
+import { PieArcDatum } from "@visx/shape/lib/shapes/Pie";
 
 import { formatNumber } from "metabase/static-viz/lib/numbers";
-import { truncateText } from "metabase/static-viz/lib/text";
-import type { ColorGetter } from "metabase/static-viz/lib/colors";
+import { ColorGetter } from "metabase/static-viz/lib/colors";
 
 import {
   START_ANGLE,
   END_ANGLE,
   CHART_WIDTH,
   GAUGE_OUTER_RADIUS,
-  CHART_VERTICAL_MARGIN,
   GAUGE_INNER_RADIUS,
-  SEGMENT_LABEL_MARGIN,
-  DISTANCE_TO_MIDDLE_LABEL_ANCHOR,
-  MAX_SEGMENT_VALUE_WIDTH,
   SEGMENT_LABEL_FONT_SIZE,
   CHART_HEIGHT,
 } from "./constants";
 import {
   limit,
-  removeDuplicateElements,
   calculateChartScale,
   calculateValueFontSize,
   calculateRelativeValueAngle,
-  calculateSegmentLabelPosition,
-  calculateSegmentLabelTextAnchor,
   getCirclePositionInSvgCoordinate,
   gaugeAccessor,
   gaugeSorter,
@@ -38,25 +29,25 @@ import {
 import GaugeNeedle from "./GaugeNeedle";
 import GaugeLabel from "./GaugeLabel";
 
-import type { Card, Data, GaugeLabelData, GaugeSegment } from "./types";
+import type { GaugeLabelData, GaugeSegment, Position } from "./types";
 
 interface GaugeProps {
-  card: Card;
-  data: Data;
+  value: number;
+  segments: GaugeSegment[];
+  gaugeLabels: GaugeLabelData[];
+  center: Position;
   getColor: ColorGetter;
 }
 
-export default function Gauge({ card, data, getColor }: GaugeProps) {
-  const settings = card.visualization_settings;
-  const segmentData = settings["gauge.segments"];
-
-  const gaugeCenterX = CHART_WIDTH / 2;
-  const gaugeCenterY = GAUGE_OUTER_RADIUS + CHART_VERTICAL_MARGIN;
-
-  const segmentMinValue = segmentData[0].min;
-  const segmentMaxValue = segmentData[segmentData.length - 1].max;
-
-  const value = data.rows[0][0];
+export default function Gauge({
+  value,
+  segments,
+  gaugeLabels,
+  center,
+  getColor,
+}: GaugeProps) {
+  const segmentMinValue = segments[0].min;
+  const segmentMaxValue = segments[segments.length - 1].max;
   const gaugeNeedleAngle = limit(
     START_ANGLE +
       calculateRelativeValueAngle(value, segmentMinValue, segmentMaxValue),
@@ -73,81 +64,6 @@ export default function Gauge({ card, data, getColor }: GaugeProps) {
     formattedValue,
     GAUGE_INNER_RADIUS,
   );
-
-  const segmentMinMaxLabels: GaugeLabelData[] = segmentData
-    .flatMap(segmentDatum => {
-      return [segmentDatum.min, segmentDatum.max];
-    })
-    // gauge segments could be continuous i.e. the current max and the next min is the same value.
-    // So we should remove duplicate elements.
-    .reduce(removeDuplicateElements, [])
-    .map((segmentValue, index, segmentValues): GaugeLabelData => {
-      const isMinSegmentValue = index === 0;
-      const isMaxSegmentValue = index === segmentValues.length - 1;
-      const segmentValueAngle =
-        START_ANGLE +
-        calculateRelativeValueAngle(
-          segmentValue,
-          segmentMinValue,
-          segmentMaxValue,
-        );
-
-      if (isMinSegmentValue) {
-        return {
-          position: [
-            -(GAUGE_INNER_RADIUS + GAUGE_OUTER_RADIUS) / 2,
-            SEGMENT_LABEL_MARGIN + DISTANCE_TO_MIDDLE_LABEL_ANCHOR,
-          ],
-          textAnchor: "middle",
-          value: formatNumber(segmentValue),
-          color: getColor("text-medium"),
-        };
-      }
-
-      if (isMaxSegmentValue) {
-        return {
-          position: [
-            (GAUGE_INNER_RADIUS + GAUGE_OUTER_RADIUS) / 2,
-            SEGMENT_LABEL_MARGIN + DISTANCE_TO_MIDDLE_LABEL_ANCHOR,
-          ],
-          textAnchor: "middle",
-          value: formatNumber(segmentValue),
-          color: getColor("text-medium"),
-        };
-      }
-
-      return {
-        position: calculateSegmentLabelPosition(segmentValueAngle),
-        textAnchor: calculateSegmentLabelTextAnchor(segmentValueAngle),
-        value: formatNumber(segmentValue),
-        color: getColor("text-medium"),
-      };
-    });
-
-  const segmentLabels: GaugeLabelData[] = segmentData
-    .filter(segment => segment.label)
-    .map((segment): GaugeLabelData => {
-      const angle =
-        START_ANGLE +
-        calculateRelativeValueAngle(
-          (segment.max + segment.min) / 2,
-          segmentMinValue,
-          segmentMaxValue,
-        );
-
-      return {
-        position: calculateSegmentLabelPosition(angle),
-        textAnchor: calculateSegmentLabelTextAnchor(angle),
-        value: truncateText(
-          segment.label,
-          MAX_SEGMENT_VALUE_WIDTH,
-          SEGMENT_LABEL_FONT_SIZE,
-        ),
-        color: getColor("text-dark"),
-      };
-    });
-
-  const gaugeLabels = segmentMinMaxLabels.concat(segmentLabels);
   const outlineColor = getColor("white");
 
   return (
@@ -161,9 +77,9 @@ export default function Gauge({ card, data, getColor }: GaugeProps) {
           transform={`scale(${calculateChartScale(gaugeLabels)})
                       translate(${-CHART_WIDTH / 2}, ${-CHART_HEIGHT / 2})`}
         >
-          <Group top={gaugeCenterY} left={gaugeCenterX}>
+          <Group top={center[1]} left={center[0]}>
             <Pie
-              data={segmentData}
+              data={segments}
               outerRadius={GAUGE_OUTER_RADIUS}
               innerRadius={GAUGE_INNER_RADIUS}
               pieValue={gaugeAccessor}
