+++ frontend/src/metabase/static-viz/components/Gauge/index.tsx
@@ -1,1 +1,1 @@
-export { default } from "./Gauge";
+export { default } from "./GaugeContainer";
